requirements.txt
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.0.4
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.13.1
chardet==3.0.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
httplib2==0.9.2
idna==2.6
isort==4.3.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
mccabe==0.6.1
mysqlclient==1.4.4
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.1.0
wadllib==1.3.2
wrapt==1.10.11
xkit==0.0.0
zope.interface==4.3.2
